refactored stuff to BaseModel, model_predict->predict_proba
