+++ python-client/giskard/models/base/autoserializable.py
@@ -49,77 +49,6 @@ class AutoSerializableModel(BaseModel, ABC):
         super().__init__(model_type, name, feature_names, classification_threshold, classification_labels)
         self.model = model
 
-    def _postprocess(self, raw_predictions):
-
-        # Convert predictions to numpy array
-        raw_predictions = self._convert_to_numpy(raw_predictions)
-
-        # We try to automatically fix issues in the output shape
-        raw_predictions = self._possibly_fix_predictions_shape(raw_predictions)
-
-        return raw_predictions
-
-    @configured_validate_arguments
-    def predict_df(self, df: pd.DataFrame):
-        raw_prediction = self.model_predict(df)
-        raw_prediction = self._postprocess(raw_prediction)
-
-        return raw_prediction
-
-    def _convert_to_numpy(self, raw_predictions):
-        return np.asarray(raw_predictions)
-
-    def _possibly_fix_predictions_shape(self, raw_predictions):
-        if not self.is_classification:
-            return raw_predictions
-
-        # Ensure this is 2-dimensional
-        if raw_predictions.ndim <= 1:
-            raw_predictions = raw_predictions.reshape(-1, 1)
-
-        # Fix possible extra dimensions (e.g. batch dimension which was not squeezed)
-        if raw_predictions.ndim > 2:
-            logger.warning(
-                f"\nThe output of your model has shape {raw_predictions.shape}, but we expect a shape (n_entries, n_classes). \n"
-                "We will attempt to automatically reshape the output to match this format, please check that the results are consistent.",
-                exc_info=True,
-            )
-
-            raw_predictions = raw_predictions.squeeze(tuple(range(1, raw_predictions.ndim - 1)))
-
-            if raw_predictions.ndim > 2:
-                raise ValueError(
-                    f"The output of your model has shape {raw_predictions.shape}, but we expect it to be (n_entries, n_classes)."
-                )
-
-        # E.g. for binary classification, prediction should be of the form `(p, 1 - p)`.
-        # If a binary classifier returns a single prediction `p`, we try to infer the second class
-        # prediction as `1 - p`.
-        if self.is_binary_classification and raw_predictions.shape[-1] == 1:
-            logger.warning(
-                f"\nYour binary classification model prediction is of the shape {raw_predictions.shape}. \n"
-                f"In Giskard we expect the shape {(raw_predictions.shape[0], 2)} for binary classification models. \n"
-                "We automatically inferred the second class prediction but please make sure that \n"
-                "the probability output of your model corresponds to the first label of the \n"
-                f"classification_labels ({self.meta.classification_labels}) you provided us with.",
-                exc_info=True,
-            )
-
-            raw_predictions = np.append(raw_predictions, 1 - raw_predictions, axis=1)
-
-        # For classification models, the last dimension must be equal to the number of classes
-        if raw_predictions.shape[-1] != len(self.meta.classification_labels):
-            raise ValueError(
-                f"The output of your model has shape {raw_predictions.shape}, but we expect it to be (n_entries, n_classes), \n"
-                f"where `n_classes` is the number of classes in your model output ({len(self.meta.classification_labels)} in this case)."
-            )
-
-        return raw_predictions
-
-    @abstractmethod
-    def model_predict(self, df):
-        ...
-
     def save(self, local_path: Union[str, Path]) -> None:
         giskard_class = infer_ml_library(self.model)
         self.meta.loader_class = giskard_class.__name__
@@ -173,23 +102,3 @@ class AutoSerializableModel(BaseModel, ABC):
                 f"Cannot load model with cloudpickle, "
                 f"{model_path} file not found and 'load_model' method isn't overriden"
             )
-
-    @classmethod
-    def load_data_preprocessing_function(cls, local_path: Union[str, Path]):
-        local_path = Path(local_path)
-        file_path = local_path / "giskard-data-preprocessing-function.pkl"
-        if file_path.exists():
-            with open(file_path, "rb") as f:
-                return cloudpickle.load(f)
-        else:
-            return None
-
-    @classmethod
-    def load_model_postprocessing_function(cls, local_path: Union[str, Path]):
-        local_path = Path(local_path)
-        file_path = local_path / "giskard-model-postprocessing-function.pkl"
-        if file_path.exists():
-            with open(file_path, "rb") as f:
-                return cloudpickle.load(f)
-        else:
-            return None
