+++ python-client/giskard/models/huggingface/__init__.py
@@ -111,7 +111,7 @@ class HuggingFaceModel(WrapperModel):
     def save_model(self, local_path):
         self.model.save_pretrained(local_path)
 
-    def model_predict(self, data):
+    def predict_proba(self, data):
         predictions = self._get_predictions(data)
 
         if self.is_classification and hasattr(predictions, "logits"):
