+++ python-client/giskard/models/tensorflow/__init__.py
@@ -40,5 +40,5 @@ class TensorFlowModel(MLFlowBasedModel):
                                      path=local_path,
                                      mlflow_model=mlflow_meta)
 
-    def model_predict(self, data):
+    def predict_proba(self, data):
         return self.model.predict(data)
