+++ python-client/tests/huggingface/test_email_classification_bert_custom_model.py
@@ -116,7 +116,7 @@ def test_email_classification_bert_custom_model():
     class tiny_bert_HuggingFaceModel(HuggingFaceModel):
         should_save_model_class = True
 
-        def model_predict(self, data):
+        def predict_proba(self, data):
             with torch.no_grad():
                 predictions = self.model(**data).logits
             return predictions.detach().cpu().numpy()
